rebase and cleanup
