+++ pg_search/src/api/operator/text.rs
@@ -16,10 +16,11 @@
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 use crate::api::operator::{
-    anyelement_text_opoid, anyelement_text_procoid, attname_from_var, estimate_selectivity,
-    find_var_relation, make_search_query_input_opexpr_node, ReturnedNodePointer,
+    anyelement_text_opoid, anyelement_text_procoid, estimate_selectivity,
+    make_search_query_input_opexpr_node, ReturnedNodePointer,
 };
 use crate::postgres::utils::locate_bm25_index;
+use crate::postgres::var::{fieldname_from_var, find_var_relation};
 use crate::query::SearchQueryInput;
 use crate::{nodecast, UNKNOWN_SELECTIVITY};
 use pgrx::{pg_extern, pg_sys, AnyElement, FromDatum, Internal, PgList};
@@ -68,7 +69,7 @@ fn text_support_request_simplify(arg: Internal) -> Option<ReturnedNodePointer> {
                 None,
                 Some((
                     rhs,
-                    attname_from_var(heaprelid, var, varattno)
+                    fieldname_from_var(heaprelid, var, varattno)
                         .expect("should be able to determine Var name"),
                 )),
             )
@@ -127,7 +128,7 @@ unsafe fn make_query_from_var_and_const(
     const_: *mut pg_sys::Const,
 ) -> (pg_sys::Oid, SearchQueryInput) {
     let (heaprelid, varattno, _) = find_var_relation(var, root);
-    let attname = attname_from_var(heaprelid, var, varattno);
+    let attname = fieldname_from_var(heaprelid, var, varattno);
     // the query comes from the rhs of the @@@ operator.  we've already proved it's a `pg_sys::Const` node
     let query_string = String::from_datum((*const_).constvalue, (*const_).constisnull)
         .expect("query must not be NULL");
