+++ pg_search/src/postgres/customscan/pdbscan/exec_methods/fast_fields/mod.rs
@@ -19,8 +19,7 @@ pub mod mixed;
 pub mod numeric;
 pub mod string;
 
-use std::rc::Rc;
-
+use crate::api::index::FieldName;
 use crate::api::HashSet;
 use crate::gucs;
 use crate::index::fast_fields_helper::{FFHelper, FastFieldType, WhichFastField};
@@ -36,6 +35,7 @@ use crate::schema::SearchIndexSchema;
 use itertools::Itertools;
 use pgrx::pg_sys::CustomScanState;
 use pgrx::{pg_sys, IntoDatum, PgList, PgOid, PgRelation, PgTupleDesc};
+use std::rc::Rc;
 use tantivy::columnar::StrColumn;
 use tantivy::termdict::TermOrdinal;
 use tantivy::DocAddress;
@@ -223,8 +223,7 @@ fn collect_fast_field_try_for_attno(
 
             // Get attribute info - use if let to handle missing attributes gracefully
             if let Some(att) = tupdesc.get((attno - 1) as usize) {
-                let att_name = att.name().to_string();
-                if schema.is_fast_field(att.name()) {
+                if schema.is_fast_field(&FieldName::from(att.name())) {
                     let ff_type = if att.type_oid().value() == pg_sys::TEXTOID
                         || att.type_oid().value() == pg_sys::VARCHAROID
                     {
@@ -232,7 +231,7 @@ fn collect_fast_field_try_for_attno(
                     } else {
                         FastFieldType::Numeric
                     };
-                    matches.push(WhichFastField::Named(att_name, ff_type));
+                    matches.push(WhichFastField::Named(att.name().to_string(), ff_type));
                 }
             }
             // If the attribute doesn't exist in this relation, just continue
