+++ pg_search/src/postgres/customscan/pdbscan/qual_inspect.rs
@@ -458,7 +458,7 @@ pub unsafe fn extract_quals(
         pg_sys::NodeTag::T_NullTest => {
             let nulltest = nodecast!(NullTest, T_NullTest, node)?;
             if let Some(field) = PushdownField::try_new(root, (*nulltest).arg.cast(), schema) {
-                if schema.is_fast_field(field.attname()) {
+                if schema.is_fast_field(&field.attname()) {
                     if (*nulltest).nulltesttype == pg_sys::NullTestType::IS_NOT_NULL {
                         Some(Qual::PushdownIsNotNull { field })
                     } else {
