+++ pg_search/src/postgres/mod.rs
@@ -43,6 +43,7 @@ mod parallel;
 pub mod storage;
 pub mod types;
 pub mod utils;
+pub mod var;
 pub mod visibility_checker;
 
 #[repr(u16)] // b/c that's what [`pg_sys::StrategyNumber`] is
